Bind specials for `seek-plans-stack`.
Any time `seek-plans-stack` is called, it must be done in a context where a large number of special variables for counting (for assessing performance) are bound.  Establish `bind-count-variables` macro for this purpose.
Note that it is the *caller's* job to handle this -- it's already done by `find-plans-stack`, but any new call direct to `seek-plans-stack` will need to be wrapped.